Project/requirements.txt
attrs==23.1.0
boto3==1.29.3
botocore==1.32.3
certifi==2023.7.22
charset-normalizer==3.3.2
cloudpickle==2.2.1
colorama==0.4.6
contextlib2==21.6.0
Cython==3.0.5
dill==0.3.7
emoji==1.2.0
exceptiongroup==1.1.3
filelock==3.13.1
fsspec==2023.10.0
gluonnlp==0.8.0
google-pasta==0.2.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwipiepy==0.16.1
kiwipiepy-model==0.16.0
kobert-tokenizer @ git+https://github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3#subdirectory=kobert_hf
kss==4.5.4
MarkupSafe==2.1.3
mpmath==1.3.0
multiprocess==0.70.15
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pathos==0.3.1
pecab==1.0.8
platformdirs==4.0.0
pluggy==1.3.0
pox==0.3.3
ppft==1.7.6.7
protobuf==4.25.1
psutil==5.9.6
py-hanspell @ git+https://github.com/ssut/py-hanspell.git@fdc6ca50c19f1c85971437a072d89d4e5ce024b8
pyarrow==14.0.1
PyMySQL==1.1.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rpds-py==0.13.1
s3transfer==0.7.0
safetensors==0.4.0
sagemaker==2.197.0
schema==0.7.5
scikit-learn==1.3.2
scipy==1.11.4
sentencepiece==0.1.99
six==1.16.0
smdebug-rulesconfig==1.0.1
soynlp==0.0.493
sympy==1.12
tblib==1.7.0
threadpoolctl==3.2.0
tokenizers==0.14.1
tomli==2.0.1
torch==2.1.0
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
zipp==3.17.0
